+++ src/agdb/db/db_context.rs
@@ -1,0 +1,3 @@
+pub struct Context {
+    pub index: i64,
+}
