+++ src/agdb/db/db_element.rs
@@ -2,7 +2,7 @@ use super::db_key_value::DbKeyValue;
 
 #[derive(Debug, PartialEq)]
 pub struct DbElement {
-    pub index: u64,
+    pub index: i64,
     pub values: Vec<DbKeyValue>,
 }
